deployment/requirements.txt
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
transformers==4.47.1
openai-whisper==20240930
opencv-python==4.10.0.84
numpy==1.26.4
nvgpu==0.10.0
soundfile==0.12.1
ffmpeg-python==0.2.0